+++ test/fixtures/frameworks/foo/lifecycle.ts
@@ -1,0 +1,41 @@
+import { Server } from 'http';
+import http from 'http';
+import { ApplicationExtension, ApplicationHook, WithApplication } from '../../../../src/decorator';
+import { ApplicationLifecycle } from '../../../../src/types';
+import { Context, Input } from '@artus/pipeline';
+import { ArtusApplication } from '../../../../src';
+
+export let server: Server;
+
+@ApplicationExtension()
+export class ApplicationHookExtension implements ApplicationLifecycle {
+  app: ArtusApplication;
+
+  constructor(@WithApplication() app: ArtusApplication) {
+    this.app = app;
+  }
+
+  @ApplicationHook()
+  async didLoad() {
+    this.app.trigger.use(async (ctx: Context) => {
+      const { data } = ctx.output;
+      data.content = { title: 'Hello Artus' };
+    });
+  }
+
+  @ApplicationHook()
+  willReady() {
+    server = http
+      .createServer(async (req: http.IncomingMessage, res: http.ServerResponse) => {
+        const input = new Input();
+        input.params = { req, res };
+        await this.app.trigger.startPipeline(input);
+      })
+      .listen(3001)
+  }
+
+  @ApplicationHook()
+  beforeClose() {
+    server?.close();
+  }
+}
